Improved support for WAV files bigger than 4 GB which only contain a single DATA chunk and a wrong length field for this DATA chunk (e.g. eac3to creates such files).

+++ src/input/r_wav.h
@@ -61,7 +61,7 @@ typedef counted_ptr<wav_demuxer_c> wav_demuxer_cptr;
 struct wav_chunk_t {
   int64_t pos;
   char id[4];
-  uint32_t len;
+  int64_t len;
 };
 
 class wav_reader_c: public generic_reader_c {
